Рефакторинг.
 - доработал адаптер
 - включил viewBinding
 - переместил некоторые классы по пакетам
 - поправил нейминг
