--- app/src/main/java/com/example/viewcomponents/Model/Model.kt
@@ -1,29 +1,0 @@
-package com.example.viewcomponents.Model
-
-class Model(idTicket: Int,
-            departureTime: String,
-            arrivalTime: String,
-            sendingPlace: String,
-            destinationPlace: String,
-            flightNumber: String,
-            flightTime: String,
-            flightDate: String
-            ) {
-
-    // id Билета
-    val idTicket: Int = 0
-    // время вылета
-    val departureTime: String = ""
-    // время прилета
-    var arrivalTime: String = ""
-    // место отправки
-    val sendingPlace: String = ""
-    // место назначения
-    val destinationPlace: String = ""
-    // номер рейса
-    val flightNumber: String = ""
-    // время полета
-    val flightTime: String = ""
-    // дата
-    val flightDate: String = ""
-}
