--- app/src/main/java/com/example/viewcomponents/data/Tickets.kt
@@ -1,67 +1,0 @@
-package com.example.viewcomponents.data
-
-import com.example.viewcomponents.Model.Model
-
-public val Tickets = listOf(
-        Model(
-                idTicket = 0,
-                departureTime = "10:00",
-                arrivalTime = "14:00",
-                sendingPlace = "Chek Lap Kok International Airport",
-                destinationPlace = "San Francisco International Airport",
-                flightNumber = "FJ128",
-                flightTime = "10 Sep",
-                flightDate = "2:00"
-        ),
-        Model(
-                idTicket = 1,
-                departureTime = "11:00",
-                arrivalTime = "15:00",
-                sendingPlace = "Chek Lap Kok International Airport",
-                destinationPlace = "Charles M. Schulz–Sonoma County Airport",
-                flightNumber = "FJ9H8",
-                flightTime = "11 Sep",
-                flightDate = "3:00"
-
-        ),
-        Model(
-                idTicket = 2,
-                departureTime = "12:00",
-                arrivalTime = "17:00",
-                sendingPlace = "Chek Lap Kok International Airport",
-                destinationPlace = "Oakland International Airport",
-                flightNumber = "HJ028",
-                flightTime = "13 Sep",
-                flightDate = "4:30"
-        ),
-        Model(
-                idTicket = 3,
-                departureTime = "15:00",
-                arrivalTime = "20:00",
-                sendingPlace = "Chek Lap Kok International Airport",
-                destinationPlace = "Norman Y. Mineta San José International Airport",
-                flightNumber = "FJ100",
-                flightTime = "14 Sep",
-                flightDate = "5:00"
-        ),
-        Model(
-                idTicket = 4,
-                departureTime = "16:30",
-                arrivalTime = "20:30",
-                sendingPlace = "Chek Lap Kok International Airport",
-                destinationPlace = "Schwechat International",
-                flightNumber = "RY168",
-                flightTime = "1 Oct",
-                flightDate = "6:00"
-        ),
-        Model(
-                idTicket = 5,
-                departureTime = "20:00",
-                arrivalTime = "01:00",
-                sendingPlace = "Chek Lap Kok International Airport",
-                destinationPlace = "Glasgow International",
-                flightNumber = "5D7RT",
-                flightTime = "9 Oct",
-                flightDate = "2:00"
-        )
-)
